requirements.txt
aiohttp==3.7.4.post0
alibabacloud-credentials==0.2.0
alibabacloud-dingtalk==1.1.94
alibabacloud-endpoint-util==0.0.3
alibabacloud-openapi-util==0.1.5
alibabacloud-tea==0.2.8
alibabacloud-tea-openapi==0.2.7
alibabacloud-tea-util==0.3.5
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
dingtalk-sdk==1.3.8
DingtalkChatbot==1.5.3
Django==3.2.8
django-cors-headers==3.10.0
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
multidict==5.2.0
mysqlclient==2.0.3
optionaldict==0.1.2
packaging==21.2
Pillow==8.4.0
pycparser==2.20
pycrypto==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.3
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
yarl==1.7.0
